@@ -6,43 +6,43 @@
   "author": "Ian Spence",
   "license": "Apache-2.0",
   "devDependencies": {
-    "@types/bootstrap": "^5.1.9",
-    "@types/prop-types": "^15.7.4",
-    "@types/react": "^17.0.42",
-    "@types/react-dom": "^17.0.14",
+    "@types/bootstrap": "^5.1.10",
+    "@types/prop-types": "^15.7.5",
+    "@types/react": "^18.0.8",
+    "@types/react-dom": "^18.0.3",
     "@types/react-router-dom": "^5.3.3",
-    "@typescript-eslint/eslint-plugin": "^5.16.0",
-    "@typescript-eslint/parser": "^5.16.0",
-    "autoprefixer": "^10.4.4",
+    "@typescript-eslint/eslint-plugin": "^5.22.0",
+    "@typescript-eslint/parser": "^5.22.0",
+    "autoprefixer": "^10.4.7",
     "copy-webpack-plugin": "^10.2.4",
     "css-loader": "^6.7.1",
-    "eslint": "^8.11.0",
+    "eslint": "^8.14.0",
     "eslint-plugin-react": "^7.29.4",
     "eslint-webpack-plugin": "^3.1.1",
     "html-loader": "^3.1.0",
     "html-webpack-plugin": "^5.5.0",
-    "sass": "^1.49.9",
+    "sass": "^1.51.0",
     "sass-loader": "^12.6.0",
     "source-map-loader": "^3.0.1",
     "style-loader": "^3.3.1",
-    "ts-loader": "^9.2.8",
-    "typescript": "^4.6.2",
-    "webpack": "^5.70.0",
+    "ts-loader": "^9.3.0",
+    "typescript": "^4.6.4",
+    "webpack": "^5.72.0",
     "webpack-cli": "^4.9.2",
     "webpack-merge": "^5.8.0"
   },
   "dependencies": {
-    "@fontsource/fira-sans": "^4.5.4",
+    "@fontsource/fira-sans": "^4.5.8",
     "@fortawesome/fontawesome-svg-core": "^6.1.1",
     "@fortawesome/free-solid-svg-icons": "^6.1.1",
     "@fortawesome/react-fontawesome": "^0.1.18",
-    "@popperjs/core": "^2.11.4",
+    "@popperjs/core": "^2.11.5",
     "@react-hook/debounce": "^4.0.0",
     "@types/debounce-promise": "^3.1.4",
     "bootstrap": "^5.1.3",
     "debounce-promise": "^3.1.2",
-    "react": "^17.0.2",
-    "react-dom": "^17.0.2",
-    "react-router-dom": "^5.3.0"
+    "react-router-dom": "^5.3.1",
+    "react": "^18.1.0",
+    "react-dom": "^18.1.0"
   }
 }
