@@ -1,10 +1,14 @@
 import * as React from 'react';
 
+interface ErrorBoundaryProps {
+    children?: React.ReactNode
+}
 interface ErrorBoundaryState {
     errorDetails?: string;
+    
 }
-export class ErrorBoundary extends React.Component<unknown, ErrorBoundaryState> {
-    constructor(props: unknown) {
+export class ErrorBoundary extends React.Component<ErrorBoundaryProps, ErrorBoundaryState> {
+    constructor(props: ErrorBoundaryProps) {
         super(props);
         this.state = {};
     }
