@@ -1,6 +1,6 @@
 export interface URLParams {
-    name: string;
-    id: string;
-    query: string;
-    label: string;
+    name?: string;
+    id?: string;
+    query?: string;
+    label?: string;
 }
